refactor all "model" packages into a single model package
