+++ core/src/game/sniper_monkey/model/player/fighter/FighterFactory.java
@@ -1,4 +1,4 @@
-package game.sniper_monkey.player.fighter;
+package game.sniper_monkey.model.player.fighter;
 
 /**
  * @author Elias Falk
