+++ core/src/game/sniper_monkey/model/world/ITimerObserver.java
@@ -1,4 +1,4 @@
-package game.sniper_monkey.world;
+package game.sniper_monkey.model.world;
 
 public interface ITimerObserver {
     void onTimerChange(int time);
