+++ core/src/game/sniper_monkey/view/player/fighter/EvilWizardView.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.math.Vector2;
-import game.sniper_monkey.player.FighterAnimation;
-import game.sniper_monkey.player.Player;
+import game.sniper_monkey.model.player.FighterAnimation;
+import game.sniper_monkey.model.player.Player;
 import game.sniper_monkey.utils.time.Time;
 import game.sniper_monkey.utils.view.AnimationUtils;
 import game.sniper_monkey.utils.view.SpriteUtils;
