+++ core/src/test/sniper_monkey/ConfigTest.java
@@ -1,13 +1,11 @@
 package sniper_monkey;
 
 import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.backends.headless.HeadlessApplication;
 import com.badlogic.gdx.backends.headless.HeadlessApplicationConfiguration;
-import game.sniper_monkey.Config;
+import game.sniper_monkey.model.Config;
 import static org.junit.Assert.*;
 
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
