+++ core/src/test/sniper_monkey/collision/CollisionPairTest.java
@@ -1,9 +1,9 @@
 package sniper_monkey.collision;
 
 import com.badlogic.gdx.math.Vector2;
-import game.sniper_monkey.collision.CollisionPair;
-import game.sniper_monkey.collision.Hitbox;
-import game.sniper_monkey.world.GameObject;
+import game.sniper_monkey.model.collision.CollisionPair;
+import game.sniper_monkey.model.collision.Hitbox;
+import game.sniper_monkey.model.world.GameObject;
 import org.junit.Before;
 import org.junit.Test;
