+++ core/src/test/sniper_monkey/collision/SpatialHashTest.java
@@ -1,16 +1,13 @@
 package sniper_monkey.collision;
 
 import com.badlogic.gdx.math.Vector2;
-import game.sniper_monkey.collision.CollisionPair;
-import game.sniper_monkey.collision.Hitbox;
-import game.sniper_monkey.collision.SpatialHash;
-import game.sniper_monkey.platform.Platform;
-import game.sniper_monkey.player.PlayerFactory;
-import game.sniper_monkey.world.GameObject;
+import game.sniper_monkey.model.collision.CollisionPair;
+import game.sniper_monkey.model.collision.SpatialHash;
+import game.sniper_monkey.model.platform.Platform;
+import game.sniper_monkey.model.world.GameObject;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
