+++ core/src/test/sniper_monkey/player/PlayerTest.java
@@ -3,10 +3,9 @@ package sniper_monkey.player;
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.backends.headless.HeadlessApplication;
 import com.badlogic.gdx.backends.headless.HeadlessApplicationConfiguration;
-import com.badlogic.gdx.graphics.g2d.Sprite;
-import game.sniper_monkey.player.Player;
-import game.sniper_monkey.player.PlayerFactory;
-import game.sniper_monkey.player.PlayerInputAction;
+import game.sniper_monkey.model.player.Player;
+import game.sniper_monkey.model.player.PlayerFactory;
+import game.sniper_monkey.model.player.PlayerInputAction;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
