+++ core/src/test/sniper_monkey/player/fighter/FighterTest.java
@@ -1,18 +1,9 @@
 package sniper_monkey.player.fighter;
 
-import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.math.Vector2;
-import game.sniper_monkey.collision.Hitbox;
-import game.sniper_monkey.player.Player;
-import game.sniper_monkey.player.fighter.Fighter;
-import game.sniper_monkey.player.fighter.FighterFactory;
-import game.sniper_monkey.world.GameObject;
-import game.sniper_monkey.world.World;
-import org.junit.Before;
+import game.sniper_monkey.model.player.fighter.Fighter;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import static org.junit.Assert.*;
-
 import static org.junit.Assert.*;
 
 public class FighterTest {
